+++ src/components/Footer/Footer.test.tsx
@@ -11,10 +11,11 @@ describe('Footer component', () => {
   it('renders the USSF portal header', () => {
     render(<Footer />)
 
-    expect(
-      screen.getByRole('img', { name: 'United States Space Force Logo' })
-    ).toHaveAttribute('alt', 'United States Space Force Logo')
-    expect(screen.getAllByRole('link')).toHaveLength(16)
+    expect(screen.getByRole('img', { name: 'USSF Portal' })).toHaveAttribute(
+      'alt',
+      'USSF Portal'
+    )
+    expect(screen.getAllByRole('link')).toHaveLength(15)
   })
 
   it('has no a11y violations', async () => {